+++ Kafka.DotNet.ksqlDb/KSql/Linq/QbservableExtensions.cs
@@ -271,7 +271,6 @@ namespace Kafka.DotNet.ksqlDB.KSql.Linq
     
     #region ObserveOn
 
-    //TODO:
     /// <summary>
     /// Wraps the source sequence in order to run its observer callbacks on the specified scheduler.
     /// </summary>
@@ -280,10 +279,6 @@ namespace Kafka.DotNet.ksqlDB.KSql.Linq
     /// <param name="scheduler">Scheduler to notify observers on.</param>
     /// <returns>The source sequence whose observations happen on the specified scheduler.</returns>
     /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="scheduler"/> is null.</exception>
-    /// <remarks>
-    /// This only invokes observer callbacks on a scheduler. In case the subscription and/or unsubscription actions have side-effects
-    /// that require to be run on a scheduler, use <see cref="Observable.SubscribeOn{TSource}(IObservable{TSource}, IScheduler)"/>.
-    /// </remarks>
     public static IQbservable<TSource> ObserveOn<TSource>(this IQbservable<TSource> source, IScheduler scheduler)
     {
       if (source == null)
@@ -294,6 +289,32 @@ namespace Kafka.DotNet.ksqlDB.KSql.Linq
 
       if (source is KStreamSet streamSet)
         streamSet.ObserveOnScheduler = scheduler;
+
+      return source;
+    }
+
+    #endregion
+
+    #region SubscribeOn
+
+    /// <summary>
+    /// Wraps the source sequence in order to run its subscription on the specified scheduler.
+    /// </summary>
+    /// <typeparam name="TSource">The type of the elements in the source sequence.</typeparam>
+    /// <param name="source">Source sequence.</param>
+    /// <param name="scheduler">Scheduler to perform subscription actions on.</param>
+    /// <returns>The original source.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="scheduler"/> is null.</exception>
+    public static IQbservable<TSource> SubscribeOn<TSource>(this IQbservable<TSource> source, IScheduler scheduler)
+    {
+      if (source == null)
+        throw new ArgumentNullException(nameof(source));
+
+      if (scheduler == null)
+        throw new ArgumentNullException(nameof(scheduler));
+
+      if (source is KStreamSet streamSet)
+        streamSet.SubscribeOnScheduler = scheduler;
 
       return source;
     }