+++ Kafka.DotNet.ksqlDb/KSql/Linq/QbservableExtensions.cs
@@ -7,6 +7,7 @@ using System.Reactive.Concurrency;
 using System.Reactive.Linq;
 using System.Reflection;
 using System.Threading;
+using System.Threading.Tasks;
 using Kafka.DotNet.ksqlDB.KSql.Query;
 using Kafka.DotNet.ksqlDB.KSql.Query.Options;
 using Kafka.DotNet.ksqlDB.KSql.Query.Windows;
@@ -317,6 +318,37 @@ namespace Kafka.DotNet.ksqlDB.KSql.Linq
         streamSet.SubscribeOnScheduler = scheduler;
 
       return source;
+    }
+
+    #endregion
+
+    #region SubscribeAsync
+
+    /// <summary>
+    /// Subscribes an element handler, an exception handler, and a completion handler to an qbservable stream.
+    /// </summary>
+    /// <typeparam name="T">The type of the elements in the source stream.</typeparam>
+    /// <param name="source">Observable sequence to subscribe to.</param>
+    /// <param name="onNext">Action to invoke for each element in the qbservable stream.</param>
+    /// <param name="onError">Action to invoke upon exceptional termination of the qbservable stream.</param>
+    /// <param name="onCompleted">Action to invoke upon graceful termination of the qbservable stream.</param>
+    /// <returns>Query id.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="onNext"/> or <paramref name="onError"/> or <paramref name="onCompleted"/> is <c>null</c>.</exception>
+    public static Task<string> SubscribeAsync<T>(this IQbservable<T> source, Action<T> onNext, Action<Exception> onError, Action onCompleted, CancellationToken cancellationToken = default)
+    {
+      if (source == null)
+        throw new ArgumentNullException(nameof(source));
+
+      if (onNext == null)
+        throw new ArgumentNullException(nameof(onNext));
+
+      if (onError == null)
+        throw new ArgumentNullException(nameof(onError));
+
+      if (onCompleted == null)
+        throw new ArgumentNullException(nameof(onCompleted));
+
+      return source.SubscribeAsync(new AnonymousObserver<T>(onNext, onError, onCompleted), cancellationToken);
     }
 
     #endregion